Renamed documentation, + added a util to readxgap for quick testing of created files

+++ src/D/qtl/plugins/output/write_xgapbin.d
@@ -93,7 +93,7 @@ class BinaryWriter(Reader, XType) {
     f = File(filename,"wb");
     myWrite(b_footprint,f);
     myWrite(b_version,f);
-    uint[1] nmatrix = [ 3 ];
+    double[1] nmatrix = [ 3 ];
     myWrite(nmatrix,f);
     myWrite(b_footprint,f);
     write_matrix!(Phenotype!double)(data.phenotypes, f, MatrixType.DOUBLEMATRIX);
